// AllHandsOnBoard/AllHandsOnBoardFrontend/src/app/task-adder/tags.interface.ts
export interface Tags {
    isIt?: boolean;
    isMaths?: boolean;
    isPhysics?: boolean;
    isChemistry?: boolean;
    isElectronics?: boolean;
    isTeaching?: boolean;
    isPresentations?: boolean;
    isForeign_languages?: boolean;
    isErasmus_students?: boolean;
    isConferences?: boolean;
    isPhysical?: boolean;
    isSport?: boolean;
}